MMR ON SITE EEEEE

+++ internal/repository/announces_repo.go
@@ -3,5 +3,7 @@ package repository
 import "ssstatistics/internal/domain"
 
 func RemoveAnnounces(ids []int32) {
-	Database.Delete(&domain.CommunicationLogs{}, ids)
+	if len(ids) > 0 {
+		Database.Delete(&domain.CommunicationLogs{}, ids)
+	}
 }
